+++ libretro-common/streams/file_stream.c
@@ -237,14 +237,15 @@ void filestream_rewind(RFILE *stream)
    stream->error_flag = false;
 }
 
-ssize_t filestream_read(RFILE *stream, void *s, size_t len)
+ssize_t filestream_read(RFILE *stream, void *s, int64_t len)
 {
    int64_t output;
 
    if (filestream_read_cb != NULL)
       output = filestream_read_cb(stream->hfile, s, len);
    else
-      output = retro_vfs_file_read_impl((libretro_vfs_implementation_file*)stream->hfile, s, len);
+      output = retro_vfs_file_read_impl(
+            (libretro_vfs_implementation_file*)stream->hfile, s, len);
 
    if (output == vfs_error_return_value)
       stream->error_flag = true;
@@ -291,7 +292,7 @@ const char *filestream_get_path(RFILE *stream)
    return retro_vfs_file_get_path_impl((libretro_vfs_implementation_file*)stream->hfile);
 }
 
-ssize_t filestream_write(RFILE *stream, const void *s, size_t len)
+ssize_t filestream_write(RFILE *stream, const void *s, int64_t len)
 {
    int64_t output;
 
@@ -390,12 +391,12 @@ int filestream_read_file(const char *path, void **buf, ssize_t *len)
    if (content_buf_size < 0)
       goto error;
 
-   content_buf = malloc(content_buf_size + 1);
+   content_buf = malloc((size_t)(content_buf_size + 1));
 
    if (!content_buf)
       goto error;
 
-   ret = filestream_read(file, content_buf, content_buf_size);
+   ret = filestream_read(file, content_buf, (int64_t)content_buf_size);
    if (ret < 0)
    {
       fprintf(stderr, "Failed to read %s: %s\n", path, strerror(errno));
